+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 
 - Spell preparation changes
 - Effects for Advantage on Skill, Ability and Checks in automated effects.
+- Support v2 Journal Pages introduced in 5.1.x. DDB Importer Journal Page to help preserve some of the unique formatting for DDB entries. DDB Importer journal enhancements will now only work on these DDB Importer journals (previously they used flags to determine if they world work). A migration will be triggered to check all world journals and journals in the DDB Journals compendium when you log in as a GM with this version or higher of DDB Importer.
 
 # Next Up